refactor: /payment 최적화(hook 분리, 타입 선언 공통 관리 등)
